refactor; review feedback
